Added Pokedex Page
No specific icon rn

The page shows a horizontal list of regions, tapping which, user can get the list of that region's pokemons. If guessed correctly, the mon should be a black silouette while if caught the image of pokemon would be normal, else there would only the id of that mon.