Fixed 'Failed to fetch pokemon species data' bug
With pokemon variants with ids 10000+, we weren't able to fetch image and were getting an error.

Issue: This was because the getPokemonImage method was also requesting pokemon data when loading the image, for which it used the id to call fetchPokemonData method, now called fetchPokemonDataFromSpeciesId, but that method required species id and not the pokemon id and hence we were getting the erorr.

Solution: I created a different method called fetchPokemonwithPokemonId which uses pokemonId to get the data.

Note: Ids that I'm talking about are not the official national dex ids, but pokeapi ids, although same for the base variants, the ids are different for special variants.