Adding Varities in Pokedex
This is part 1
We are now showing even varities of the pokemon on the pokedex gridview element if the base form is not caught as of yet.

This requirs more testing. It isn't done yet.

And then we'll move to adding varities in the individual pokemon sectio which opens upon tapping a caught pokemon's image.

But first a bug I've found.